Updating build. Adding events package. Adding tests.
I decided that the best way to deal with querying events is to handle them by replaying them. Normally, events are processed at a delay or in real-time. Regardless, events are intended to be processed at the earliest possible time rather than being traversed in a database. It is probably easier than traversing events to replay them and handle them as they happen.